+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🐛 fix: bugs on duplication check fixed and documentation added for every function in manager.gd script
 
 3.0.1 (2022-12-28)
 ------------------